refactor: extract session cache and envelope modules
